[Refactor] - Adds throwable extension method to wrap network related exceptions to NetworkException
